Struct rf24::RF24

pub struct RF24 {
    handle: RF24,
}

An nRF24 device

Fields
handle: RF24
Methods

impl RF24

pub fn from_pins(ce: u16, csn: u16) -> Self

Create a nRF24 device interface from the pin numbers where the Chip Enable and Chip Select Not pins are connected.

pub fn from_spi_device(ce: u16, spi_device: u16) -> Self

Create a nRF24 device interface from the pin number where the Chip Enable is connected and its SPI device index.

pub fn from_pins_with_speed(ce: u16, csn: u16, spi_speed: SPISpeed) -> Self

Create a nRF24 device interface from the pin numbers where the Chip Enable and Chip Select Not pins are connected and the bus speed to use.

pub fn from_spi_device_with_speed(
ce: u16,
spi_device: u16,
spi_speed: SPISpeed
) -> Self

Create a nRF24 device interface from the pin number where the Chip Enable is connected, its SPI device index and the bus speed to use.

pub fn begin(&mut self) -> bool

Begin operating with the chip.

This method has to be called before any other tasks can be done.

pub fn chip_connected(&mut self) -> bool

Returns whether a nRF24 has been detected.

pub fn failure_detected(&mut self) -> bool

Returns whether a hardware failure has occurred. If it did, the driver will cease its operation by default.

pub fn listen_start(&mut self)

Signals the radio to start listening for packets.

pub fn listen_stop(&mut self)

Signals the radio to stop listening for packets, allowing the radio itself to send packets.

pub fn available(&mut self) -> bool

Returns whether there are packets available to be read.

pub fn is_p_variant(&mut self) -> bool

Returns whether the hardware is an nRF24L01+ (with power amplifier) or not.

pub fn get_power_amplifier_level(
&mut self
) -> Result<PowerAmplifierLevel, RF24Error>

Returns the current power amplifier level

pub fn set_power_amplifier_level(&mut self, pa_level: PowerAmplifierLevel)

Sets the current power amplifier level

pub fn get_data_rate(&mut self) -> Result<DataRate, RF24Error>

Returns the device's currently configured transmission rate

pub fn set_data_rate(&mut self, data_rate: DataRate) -> bool

Sets the device's transmission rate

pub fn get_crc_length(&mut self) -> Result<CRCLength, RF24Error>

Returns the device's currently configured message CRC length

pub fn set_crc_length(&mut self, crc_length: CRCLength)

Sets the device's currently configured message CRC length

pub fn disable_crc(&mut self)

Disables CRC validation

pub fn set_address_width(&mut self, addr_width: u8)

Sets the device address width used by the device

pub fn set_retries(&mut self, delay: u8, count: u8)

Sets the number of message retransmissions that may ocurr at most if acknowledgements are not received and the delay between them.

pub fn set_channel(&mut self, channel: u8)

Sets the frequency channel where the device will listen and send packets to.

This channel is represented by a number between 1 and 125 (both inclusive), channel 1 being on a radio frequency of 2400 MHz and channel 125 being on 2525 MHz, with 1 MHz of spacing between frequencies.

pub fn get_channel(&mut self) -> u8

Returns the frequency channel the device is currently on.

This channel is represented by a number between 1 and 125 (both inclusive), channel 1 being on a radio frequency of 2400 MHz and channel 125 being on 2525 MHz, with 1 MHz of spacing between frequencies.

pub fn power_down(&mut self)

Signals the device to enter low power mode, a deep sleep state where the device needs to be woken back up before resuming operations.

pub fn power_up(&mut self)

Signals the device to leave low power mode and return back to normal operation.

pub fn read(&mut self, buffer: &mut [u8]) -> Result<(), RF24Error>

Reads the contents of the next packet onto the passed buffer.

The buffer must be already zero-filled with the requested message length.

pub unsafe fn read_bytes(
&mut self,
buffer: &mut [u8],
num_bytes: u8
) -> Result<(), RF24Error>

Reads the given amount of bytes off the contents of the next packet onto the passed buffer.

The buffer must be already zero-filled with the requested message length. WARNING: Using the incorrect message length may overflow the buffer and break memory safety, hence why this method is marked unsafe.

pub fn read_dynamic_payload(
&mut self,
buffer: &mut Vec<u8>
) -> Result<bool, RF24Error>

Reads off the contents of the next dynamically-sized packet onto the passed vector.

The vector's contents are cleared and the vector is resized properly within the method.

pub fn write(&mut self, buffer: &[u8]) -> Result<bool, RF24Error>

Transmits the buffer's contents as a packet, blocking until an acknowledgement has been received for such packet (returning true) or the retries limits have been reached (returning false).

pub fn write_unreliable(&mut self, buffer: &[u8]) -> Result<bool, RF24Error>

Transmits the buffer's contents as a packet while signaling that no acknowledgement is required to be sent by the receiver.

RF24::enable_dynamic_ack must have been called once first.

pub fn write_fast(&mut self, buffer: &[u8]) -> Result<bool, RF24Error>

Transmits the buffer's contents as a packet, not blocking until all the device's internal packet buffers are full.

pub fn write_fast_unreliable(
&mut self,
buffer: &[u8]
) -> Result<bool, RF24Error>

Transmits the buffer's contents as a packet, not blocking until all the device's internal packet buffers are full, and also disabling acknowledgements for the given packet.

RF24::enable_dynamic_ack must have been called once first.

pub fn write_timeout(
&mut self,
buffer: &[u8],
timeout: u32
) -> Result<bool, RF24Error>

Transmits the buffer's contents as a packet, not blocking until all the device's internal packet buffers are full. When it blocks, it will do so until a payload has been successfully written or the specified timeout period (in milliseconds) is reached.

pub fn set_dynamic_payloads(&mut self, enable: bool)

Toggle whether dynamically-sized payload support is enabled or not.

You must call this function with a value of true before sending or receiving dynamic payloads.

pub fn do_fast_write_transaction<T>(&mut self, transaction: T) where
T: Fn(),

Taking a closure, executes its code, then calls RF24::tx_standby after the code has executed. Intented to be used to execute multiple fast writes at once safely and by specification.

pub fn tx_standby(&mut self)

Drop the radio from STANDBY-II mode to STANDBY-I mode, blocking until all packets are sent.

pub fn tx_standby_timeout(&mut self, timeout: u32, start_tx: bool) -> bool

Drop the radio from STANDBY-II mode to STANDBY-I mode, blocking until all packets are sent or the timeout (in miliseconds) is reached. If start_tx is true, the driver will also exit out of listening mode so the packets can be sent.

pub fn open_writing_pipe(&mut self, address: &[u8]) -> Result<(), RF24Error>

Open a pipe for writing packets out as the specified radio address.

pub fn open_reading_pipe(
&mut self,
pipe_number: u8,
address: &[u8]
) -> Result<(), RF24Error>

Open a pipe for reading packets in, listening for packets with the specified address.

Up to 6 pipes, identified by their pipe numbers, can be opened for simultaneous reading before starting to listen.

pub fn close_reading_pipe(&mut self, pipe_number: u8)

Closes the specified reading pipe

pub fn print_details(&mut self)

Prints detailed debugging information about the driver and the hardware state to the standard output

pub fn pipe_available(&mut self) -> Option<u8>

Returns whether there are packets available to be read, and if so, on which reading pipe they are.

pub fn is_rx_full(&mut self) -> bool

Returns whether the reception queues on the device are full, meaning read request should be dispatched ASAP, as otherwise the device may lose packets.

pub fn flush_tx(&mut self) -> u8

Empty the transmission buffer.

pub fn flush_rx(&mut self) -> u8

Empty the reception buffer.

pub fn reuse_tx(&mut self)

Signal the chip to indefinitely resend the currently-available payload in the transmission buffers until another payload is written or RF24::flush_tx is called.

pub fn set_auto_ack(&mut self, enable: bool)

Toggle automatic packet acknowledgement.

pub fn set_auto_ack_pipe(&mut self, pipe: u8, enable: bool)

Toggle automatic packet acknowledgement on the specified reading pipe.

pub fn is_ack_payload_available(&mut self) -> bool

Returns whether there is an acknowledgement packet payload available for reading.

pub fn enable_ack_payload(&mut self)

Enable writing and reading of custom payloads on acknowledgement packets.

They are dynamic payloads by nature, and thus they must be enabled. ACK payloads may work better on reading pipe 1 instead of 0.

pub fn write_ack_payload(
&mut self,
pipe_number: u8,
buffer: &[u8]
) -> Result<(), RF24Error>

Write a custom payload for the next acknowledgement packet sent on the specified reading pipe, so that the next time a packet on this pipe, the acknowledgement sent for it will contain the custom payload.

Only three custom acknowledgement payloads can be pending at any time, and dynamic payloads must be enabled before sending such a payload.

pub fn enable_dynamic_ack(&mut self)

Enable single-write basis acknowledgement adjustements.

pub fn get_payload_size(&mut self) -> u8

Returns the currently set static payload size

pub fn set_payload_size(&mut self, size: u8) -> Result<(), RF24Error>

Sets the currently set static payload size

pub fn get_dynamic_payload_size(&mut self) -> Option<u8>

Get the size in bytes of the next dynamic payload. If None, it was a corrupt payload, and it has been automatically flushed.

pub fn inspect_interrupt(&mut self) -> InterruptData

When an interrupt is received, this function can be called to identify the cause of the interrupt.

pub fn mask_interrupts(&mut self, mask: InterruptData)

Masks the given interrupt triggers set to true from generating interrupts.

pub fn carrier_available(&mut self) -> bool

Returns whether there was a carrier on the current frequency channel for the previous listening period.

Can be used to check for interference.

pub fn test_rpd(&mut self) -> bool

Returns whether there was a signal greater than -64dBm on the current frequency channel for the previous listening period. Only supported on nRF24L01P(+) chips.

Can be used to check for interference and to aid channel hopping.